requirements.txt
albumentations==1.3.1
comet_ml==3.33.4
coremltools==6.3.0
Flask==2.0.3
ipython==8.0.1
keras==2.8.0
keras_nightly==2.5.0.dev2021032900
matplotlib==3.5.1
mss==9.0.1
numpy==1.23.3
numpy==1.19.2
onnx==1.14.0
onnxruntime==1.15.0
onnxsim==0.4.31
opencv_contrib_python==4.5.5.62
opencv_python==4.6.0.66
openvino==2023.0.0
packaging==21.3
paddle==1.0.2
pafy==0.5.5
pandas==1.4.0
Pillow==9.0.1
Pillow==9.5.0
protobuf==3.19.4
psutil==5.9.4
pycocotools==2.0.6
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
scipy==1.8.0
seaborn==0.11.2
setuptools==67.3.3
setuptools==58.0.4
streamlit==1.11.1
tensorflow==2.8.0
tensorflowjs==4.7.0
tensorrt==8.6.1
tflite_runtime==2.12.0
tflite_support==0.4.3
thop==0.1.1.post2209072238
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.1
tritonclient==2.34.0
ultralytics==8.0.42
x2paddle==1.4.1
